requirements.txt
django
django-ninja
pillow
django-cors-headers